requirements.txt
pdfplumber
python-dotenv==1.0.0
google-generativeai==0.3.2
